Se crearon package para ditribuir las clases entidades y los JPACONTROLLERS.
