+++ Horizontes_Urbanos/src/main/java/Modelo/entities/Telefono_asesor.java
@@ -1,6 +1,7 @@
 
-package Modelo;
+package Modelo.entities;
 
+import Modelo.entities.Asesor;
 import java.io.Serializable;
 import javax.persistence.Column;
 import javax.persistence.Entity;
@@ -13,32 +14,32 @@ import javax.persistence.SequenceGenerator;
 import javax.persistence.Table;
 
 /**
- * Clase que determina la entidad Correo Asesor
+ * Clase que determina la entidad Telefono Asesor
  * 
  * @author Santiago
  */
 @Entity
-@Table(name = "Correo_asesor")
-public class Correo_asesor implements Serializable {
+@Table(name = "Telefono_asesor")
+public class Telefono_asesor implements Serializable {
     
     @Id
-    @GeneratedValue(strategy = GenerationType.SEQUENCE,generator ="seq_idcoase")
-    @SequenceGenerator(name = "seq_idcoase", sequenceName = "seq_idcoase", allocationSize = 1)
+    @GeneratedValue(strategy = GenerationType.SEQUENCE,generator ="seq_idTelAse")
+    @SequenceGenerator(name = "seq_idTelAse", sequenceName = "seq_idTelAse", allocationSize = 1)
     private int id;
     
     @Column(nullable = true)
-    private String correo;
+    private Integer telefono;
     
     @ManyToOne(optional=false)
     @JoinColumn(name = "Cedula_asesor", referencedColumnName = "Cedula")
     private Asesor asesor;
 
-    public Correo_asesor() {
-    } 
+    public Telefono_asesor() {
+    }
 
-    public Correo_asesor(int id, String correo, Asesor asesor) {
+    public Telefono_asesor(int id, Integer telefono, Asesor asesor) {
         this.id = id;
-        this.correo = correo;
+        this.telefono = telefono;
         this.asesor = asesor;
     }
 
@@ -50,12 +51,12 @@ public class Correo_asesor implements Serializable {
         this.id = id;
     }
 
-    public String getCorreo() {
-        return correo;
+    public Integer getTelefono() {
+        return telefono;
     }
 
-    public void setCorreo(String correo) {
-        this.correo = correo;
+    public void setTelefono(Integer telefono) {
+        this.telefono = telefono;
     }
 
     public Asesor getAsesor() {
