+++ Horizontes_Urbanos/src/main/java/Modelo/jpa_controllers/ClienteJpaController.java
@@ -1,6 +1,11 @@
 
-package Modelo;
+package Modelo.jpa_controllers;
 
+import Modelo.entities.Venta;
+import Modelo.entities.Telefono_cliente;
+import Modelo.entities.Pago;
+import Modelo.entities.Correo_cliente;
+import Modelo.entities.Cliente;
 import Modelo.exceptions.IllegalOrphanException;
 import Modelo.exceptions.NonexistentEntityException;
 import Modelo.exceptions.PreexistingEntityException;
