+++ Horizontes_Urbanos/src/main/java/Modelo/jpa_controllers/Correo_asesorJpaController.java
@@ -1,6 +1,8 @@
 
-package Modelo;
+package Modelo.jpa_controllers;
 
+import Modelo.entities.Correo_asesor;
+import Modelo.entities.Asesor;
 import Modelo.exceptions.NonexistentEntityException;
 import java.io.Serializable;
 import java.util.List;
