+++ Horizontes_Urbanos/src/main/java/Modelo/jpa_controllers/PagoJpaController.java
@@ -1,6 +1,10 @@
 
-package Modelo;
+package Modelo.jpa_controllers;
 
+import Modelo.entities.Venta;
+import Modelo.entities.Pago;
+import Modelo.entities.Cliente;
+import Modelo.entities.Asesor;
 import Modelo.exceptions.NonexistentEntityException;
 import java.io.Serializable;
 import java.util.List;
