+++ Horizontes_Urbanos/src/main/java/Modelo/jpa_controllers/Telefono_clienteJpaController.java
@@ -1,6 +1,8 @@
 
-package Modelo;
+package Modelo.jpa_controllers;
 
+import Modelo.entities.Telefono_cliente;
+import Modelo.entities.Cliente;
 import Modelo.exceptions.NonexistentEntityException;
 import java.io.Serializable;
 import java.util.List;
